Unneeded changes to IntCodeVM to change input/output handling.
By default behaviour is exactly the same as it was before, however
now we can optionally enable the vm to throw an interrupt exception
whenever there is new output, or make it just stop running if there
is no input without throwing an exception.

Mostly pointless, but I think makes some of the days a bit nicer.

All days have been updated to explicitly enable/disable interrupts
as desired.